+++ spec/controllers/integral/backend/posts_controller_spec.rb
@@ -10,9 +10,9 @@ module Integral
       let(:body) { 'foobar body.' }
       let(:description) { Faker::Lorem.paragraph(8)[0..150] }
       let(:tag_list) { 'foo,bar,tags' }
-      let(:post_params) { { title: title, body: body, description: description, tag_list: tag_list, slug: slug } }
+      let(:post_params) { { title: title, body: body, description: description, tag_list: tag_list, slug: slug, user_id: user.id } }
       let(:user) { create(:post_manager) }
-      let(:user_post) { create(:integral_post) }
+      let!(:user_post) { create(:integral_post) }
 
       describe 'GET index' do
         context 'when not logged in' do
@@ -76,6 +76,34 @@ module Integral
         end
       end
 
+      describe 'POST duplicate' do
+        context 'when not logged in' do
+          it 'redirects to login page' do
+            post :duplicate, params: { id: user_post.id }
+
+            expect(response).to redirect_to new_user_session_path
+          end
+        end
+
+        context 'when logged in' do
+          before { sign_in user }
+
+          context 'when valid post params supplied' do
+            it 'redirects to the cloned post' do
+              post :duplicate, params: { id: user_post.id }
+
+              expect(response).to redirect_to edit_backend_post_path(Integral::Post.last.id)
+            end
+
+            it 'saves a new post' do
+              expect {
+                post :duplicate, params: { id: user_post.id }
+              }.to change(Post, :count).by(1)
+            end
+          end
+        end
+      end
+
       describe 'GET new' do
         context 'when not logged in' do
           before { get :new }
@@ -86,12 +114,13 @@ module Integral
         context 'when logged in' do
           before do
             sign_in user
-            allow(Post).to receive(:new).and_return :foo
+            allow(Post).to receive(:new).and_return user_post
             get :new
           end
 
           it { expect(response.status).to eq 200 }
-          it { expect(assigns(:resource)).to eq :foo }
+          it { expect(assigns(:resource)).to eq user_post }
+          it { expect(assigns(:resource).user).to eq user }
           it { expect(response).to render_template 'new' }
         end
       end